🎨 [refactor] Changed `borehole_temperature_models` -> `icetemp`
A previous commit updated the package name to `icetemp`. This commit updates all of the source code to match the new name.
